Set Trainer Levels Dynamically
when FLAG_DYNAMIC_TRAINER_LEVELS is set, the levels of regular over-world enemy trainers will scale to be near the average party-level of the player.

In Oldale Town (at least in this commit), two signs behind the mart switch this flag. Clicking (A) on the left sign turns the dynamic_trainer_levels flag on, clicking (A) on the right sign turns it off.

Might want versions with more, or less, level-tolerance.

almost definitely also want a flag that sets all enemy mons to the same level as player's highest level mon.